+++ Echo/user-service/src/main/java/com/echoteam/app/services/UserServiceImpl.java
@@ -4,19 +4,12 @@ import com.echoteam.app.dao.UserRepository;
 import com.echoteam.app.entities.User;
 import com.echoteam.app.entities.dto.UserDTO;
 import com.echoteam.app.exceptions.ParameterIsNotValidException;
-import com.echoteam.app.exceptions.UniqueRecordAlreadyExistsException;
-import jakarta.persistence.Column;
 import jakarta.persistence.EntityNotFoundException;
 import lombok.RequiredArgsConstructor;
-import org.hibernate.exception.ConstraintViolationException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.ErrorResponse;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.lang.reflect.Field;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.Optional;
 
 @RequiredArgsConstructor
@@ -25,6 +18,7 @@ public class UserServiceImpl implements UserService{
 
     private final UserRepository userRepository;
 
+    @Transactional
     @Override
     public List<UserDTO> getAll() {
         List<User> users = userRepository.findAll();
@@ -33,6 +27,7 @@ public class UserServiceImpl implements UserService{
                 .toList();
     }
 
+    @Transactional
     @Override
     public UserDTO getById(Long id) {
         Optional<User> user = userRepository.findById(id);
@@ -43,56 +38,36 @@ public class UserServiceImpl implements UserService{
         }
     }
 
+    @Transactional
     @Override
-    public UserDTO createUser(UserDTO userDto) throws ParameterIsNotValidException,
-                                                        UniqueRecordAlreadyExistsException {
+    public UserDTO createUser(UserDTO userDto) {
         if (userDto.id() != null) {
             throw new ParameterIsNotValidException("There is no need user with id.");
         }
-        try {
-             return userRepository.save(User.of(userDto)).toDTO();
-        } catch (DataIntegrityViolationException ex) {
-            String errorMessage = recognizeDataIntegrityViolationException(ex);
-            throw new UniqueRecordAlreadyExistsException(errorMessage, ex);
-        }
+        return userRepository.save(User.of(userDto)).toDTO();
     }
 
+    @Transactional
     @Override
-    public UserDTO updateUser(UserDTO userDto) throws ParameterIsNotValidException, UniqueRecordAlreadyExistsException {
+    public UserDTO updateUser(UserDTO userDto) {
         Optional<User> optionalUser = userRepository.findById(userDto.id());
         if (optionalUser.isPresent()) {
             User dbUser = optionalUser.get();
             dbUser.acceptChanges(User.of(userDto));
-            try {
-                return userRepository.save(dbUser).toDTO();
-            } catch (DataIntegrityViolationException ex) {
-                String errorMessage = recognizeDataIntegrityViolationException(ex);
-                throw new UniqueRecordAlreadyExistsException(errorMessage, ex);
-            }
+            return userRepository.save(dbUser).toDTO();
         } else {
-            throw new EntityNotFoundException(String.format("UserDTO with id: %d not found.", userDto.id()));
+            throw new EntityNotFoundException(String.format("User with id %d not found.", userDto.id()));
         }
     }
 
+    @Transactional
     @Override
     public void deleteById(Long id) {
+        boolean exists = userRepository.existsById(id);
+        if (!exists) {
+            throw new EntityNotFoundException(String.format("User with id %d not found.", id));
+        }
         userRepository.deleteById(id);
     }
 
-    private String recognizeDataIntegrityViolationException(DataIntegrityViolationException ex) {
-        String message = ex.getMessage();
-        String requestMessage = "User with such %s already exist.";
-
-        Class userClass = User.class;
-        Field[] declaredFields = userClass.getDeclaredFields();
-        for (Field field : declaredFields) {
-            if (field.isAnnotationPresent(Column.class)) {
-                Column columnAnnotation = field.getAnnotation(Column.class);
-                if (columnAnnotation.unique() && message.contains("(" + columnAnnotation.name() + ")")) {
-                    return requestMessage.formatted(columnAnnotation.name());
-                }
-            }
-        }
-        return "Something was wrong. Try again later please.";
-    }
 }
